Team_49/requirements.txt
Flask
flask-cors
gunicorn
tensorflow==2.16.1
numpy
pillow
opencv-python
